move ape file
